backend/requirements/base.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.6,<5.0
django-filter>=24.3,<26.0
dj-database-url>=2.2,<3.0
psycopg[binary]>=3.2,<4.0
celery>=5.4,<6.0
redis>=5.0,<6.0
django-storages>=1.14,<2.0
boto3>=1.34,<2.0
python-dotenv>=1.0,<2.0
whitenoise>=6.8,<7.0
drf-spectacular>=0.28,<1.0
pypdf>=5.1,<6.0
google-auth>=2.38,<3.0
requests>=2.32,<3.0
